+++ src/game_io/game_player_ai_persistent_packet.cc
@@ -101,7 +101,7 @@ void GamePlayerAiPersistentPacket::read(FileSystem& fs, Game& game, MapObjectLoa
 				assert(player->ai_data.remaining_basic_buildings.empty());
 
 				size_t remaining_basic_buildings_size = fr.unsigned_32();
-				for (uint16_t i = 0; i < remaining_basic_buildings_size; ++i) {
+				for (size_t i = 0; i < remaining_basic_buildings_size; ++i) {
 					// Buildings saved as strings
 					const std::string building_string = fr.string();
 					const Widelands::DescriptionIndex bld_idx =
